seicom
learning and stay updated with the latest trends and advancements in the field to remain competitive in the job market.BCA (Bachelor of Computer Applications) is an undergraduate degree program in the field of computer applications. It is a three-year program typically comprising six semesters. BCA is designed to provide students with a strong foundation in computer science and its applications in various industries.